Moves plot.py functionality and deletes
