+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added 
+### Added
 - Added a setup.cfg
 - get_direct_beam_position now supports lazy proccessing (#648)
 
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Azimuthal integration has been refactored (see PR #625 for details)
 
 ### Removed
-- The local_gaussian_method for subpixel refinement has been removed
+- The local_gaussian_method for subpixel refinement
+- utils.plot, functionality now in signals.diffraction_vectors
 
 ## 2020-10-04 - version 0.12.2
 ### Added
